Reminder feature initializing
