+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_btc"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".CoinReminderActivity"
+        <activity android:name=".ChartDataActivity"
             android:parentActivityName=".MainActivity">
         </activity>
         <activity
@@ -23,6 +23,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity> <!-- Sample AdMob app ID: ca-app-pub-3940256099942544~3347511713 -->
+        <receiver android:name=".adapter.ReminderBroadcast"/>
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
             android:value="ca-app-pub-2313560120112536~7960865514" />
