--- app/src/main/java/com/abraxel/cryptocurrency/IVolley.java
@@ -1,11 +1,0 @@
-package com.abraxel.cryptocurrency;
-
-
-import com.abraxel.cryptocurrency.model.ChartData;
-import org.json.JSONObject;
-
-import java.util.List;
-
-public interface IVolley {
-    List<ChartData> onResponse(JSONObject response);
-}
