+++ app/src/main/java/com/abraxel/cryptocurrency/MethodServer.java
@@ -260,7 +260,4 @@ public class MethodServer extends MainActivity {
         return context.getResources().getIdentifier(imageName, "drawable", context.getPackageName());
     }
 
-
-
-
 }
