+++ app/src/main/java/com/abraxel/cryptocurrency/adapter/ReminderBroadcast.java
@@ -1,0 +1,24 @@
+package com.abraxel.cryptocurrency.adapter;
+
+import android.app.Notification;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.NotificationManagerCompat;
+
+import com.abraxel.cryptocurrency.R;
+
+public class ReminderBroadcast extends BroadcastReceiver {
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(context, "cryptocurrency")
+                .setSmallIcon(R.drawable.ic_baseline_money_24)
+                .setContentTitle("Kripto Para TL")
+                .setContentTitle("Ayarlanan tutar bilgisi...")
+                .setPriority(NotificationCompat.PRIORITY_HIGH);
+        NotificationManagerCompat notificationCompat = NotificationManagerCompat.from(context);
+        notificationCompat.notify(200, builder.build());
+    }
+}
