+++ app/src/main/res/layout/crypto_item_list.xml
@@ -75,6 +75,15 @@
 
             </LinearLayout>
 
+            <Button
+                android:layout_width="30dp"
+                android:layout_height="30dp"
+                android:textColor="#000000"
+                android:layout_marginTop="10dp"
+                android:layout_marginRight="10dp"
+                android:background="@drawable/baseline_notifications_none_24"
+                android:id="@+id/remind_me"/>
+
         </LinearLayout>
     </androidx.cardview.widget.CardView>
